federal_eventos/requirements2.txt

dj-database-url==0.5.0
django-extensions==3.1.1
django-utils-six==2.0
django-widget-tweaks==1.4.8
isort==5.8.0
python-decouple==3.4
ipdb